+++ docs/changelog.rst
@@ -18,6 +18,7 @@ v4.1.0 - 2020-xx-xx
 - ``Added``: Remote datalogger: Read telegrams from network socket [`#1057 <https://github.com/dennissiemensma/dsmr-reader/issues/1057>`_]
 
 - ``Changed``: Remote datalogger: Changed config to env vars [`#1057 <https://github.com/dennissiemensma/dsmr-reader/issues/1057>`_]
+- ``Changed``: Enabled retention by default for new installations [`#1000 <https://github.com/dennissiemensma/dsmr-reader/issues/1000>`_]
 
 - ``Fixed``: Polyphase detection for Fluvius meters [`#1052 <https://github.com/dennissiemensma/dsmr-reader/issues/1052>`_]